Create mailboxes that don't exist
From procmailrc(5):

If the mailbox name ends in "/", then this directory is presumed to be a
maildir folder; i.e., procmail will deliver the message to a file in a
subdirectory named "tmp" and rename it to be inside a subdirectory named
"new".  If the mailbox is specified to be an MH folder or maildir
folder, procmail will create the necessary directories if they don't
exist, rather than treat the mailbox as a non-existent filename.
Use '--' to separate paths from revisions, like this:

From RFC 5228 Sec 4.1:

If the specified mailbox doesn't exist, the implementation MAY treat it
as an error, create the mailbox, or deliver the message to an
implementation-defined mailbox.

From RFC 5490 Sec 3.2:

If the optional ":create" argument is specified with "fileinto", it
instructs the Sieve interpreter to create the specified mailbox, if
needed, before attempting to deliver the message into the specified
mailbox.  If the mailbox already exists, this argument is ignored.
Failure to create the specified mailbox is considered to be an error.

The capability string for use with the ":create" parameter is "mailbox".