Corrected procmail-related environment vars
procmail's concept of the MAILDIR variable is simply its current working
directory; it has nothing to do with any Maildir that it may or may not
end up delivering to.

Treat the initial context as immutable, so that we can rely on its
environment variables even if a procmailrc assigns to them.

We consider procmail action or forwarding destination that delivers to
~/Maildir/ as an intent to deliver to Sieve's natural INBOX.  For that
reason, we set the initial value of DEFAULT to ~/Maildir/ and compare
against that when translating filesystem paths into mailbox names.