Added crude support for procmailrc inclusions
In procmailrc, assigning to the magic INCLUDERC variable causes the
referenced procmailrc to be included.  Ideally, we would transpile the
included script separately, then call it using the Sieve include control
(RFC 6609 Sec 3.2).  For now, to avoid manipulating multiple files, we
simply inline it.

We also explicitly handle assignments to the magic HOST and SWITCHRC
variables as unsupported features.