Crude spamc command emulation
We presume that the message has already been passed through SpamAssassin by
the time the Sieve filter is executed, and that such execution of SpamAssassin
is in a global rather than a per-user context, such that it does not respect
the user's personal configuration in ~/.spamassassin/user_prefs.

Therefore, if procmail contains an invocation of spamc, the Sieve script will
instead manipulate the message's SpamAssassin-produced headers to make it look
like spam (effectively blacklisting it) or look like ham (effectively
whitelisting it).