Handling delimiters in procmailrc header regexes
In procmailrc, header conditionals are executed by applying a regex to the
entire "Header: value" string.  In contrast, Sieve header and address tests
specify the header and pattern separately.

When a procmail executes

    * ^To:.localpart@example\.org

... the dot before the localpart would match the single space that
_customarily_ appears in mail headers after the colon.  Note that RFC 2822 Sec
2.2 does not require message headers to use ": " as the delimiter between the
field name and the field body.  Rather, values that appear in the field body
are allowed to contain optional whitespace (such as atoms that constitute an
e-mail address -- see RFC 2822 Sec 3.2.3 and Sec 3.2.4).

Nevertheless, if we heuristically translate procmailrc regexes into Sieve
header tests or address tests, we shouldn't translate a procmail condition

    * ^To:.localpart@example\.org

... into

    if header :is "?localpart@example.org"

... because the question mark would set an expectation of a leading space to be
disregarded.  Strictly speaking, such a procmailrc condition, with the dot, is
poorly formulated, but it tends to work in practice because the dot matches the
space that _customarily_ appears in messages.  Nevertheless, when translating
procmail rules that people tend to write into Sieve, we should strive to guess
the author's intent and produce rules that work in practice.

Note that it is Dovecot Pigeonhole that is arguably parsing messages
improperly, by discarding leading whitespace when it parses headers.  See
https://github.com/dovecot/pigeonhole/blob/0.5.14/src/lib-sieve/sieve-message.c#L1084

    struct message_parser_settings mparser_set = {
            .hdr_flags = MESSAGE_HEADER_PARSER_FLAG_SKIP_INITIAL_LWSP,
            .flags = MESSAGE_PARSER_FLAG_INCLUDE_MULTIPART_BLOCKS,
    };

... which eventually leads to
https://github.com/dovecot/core/blob/2.3.14/src/lib-mail/message-header-parser.c#L281

    if ((ctx->flags & MESSAGE_HEADER_PARSER_FLAG_SKIP_INITIAL_LWSP) != 0) {
            /* get value. skip all LWSP after ':'. Note that
               RFC2822 doesn't say we should, but history behind
               it..
               Exception to this is if the value consists only of
               LWSP, then skip only the one LWSP after ':'. */